+++ src/main/java/com/metacoding/springv2/Springv2Application.java
@@ -2,7 +2,6 @@ package com.metacoding.springv2;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @SpringBootApplication
 public class Springv2Application {
