+++ src/main/java/com/metacoding/springv2/auth/AuthController.java
@@ -1,14 +1,18 @@
-package com.metacoding.springv2.web;
+package com.metacoding.springv2.auth;
 
-import com.metacoding.springv2.core.util.Resp;
-import com.metacoding.springv2.domain.auth.*;
-import com.metacoding.springv2.domain.user.*;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
+import com.metacoding.springv2.core.util.Resp;
+import com.metacoding.springv2.user.UserService;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @RestController
