+++ src/main/java/com/metacoding/springv2/board/BoardResponse.java
@@ -1,0 +1,56 @@
+package com.metacoding.springv2.board;
+
+import java.util.List;
+
+import com.metacoding.springv2.reply.Reply;
+
+public class BoardResponse {
+
+    public record DTO(
+            Integer id,
+            String title,
+            String content) {
+        public DTO(Board board) {
+            this(
+                    board.getId(),
+                    board.getTitle(),
+                    board.getContent());
+        }
+    }
+
+    public record DetailDTO(
+            Integer boardId,
+            String title,
+            String content,
+            Integer userId,
+            String username,
+            Boolean isOwner,
+            List<ReplyDTO> replies) {
+        public DetailDTO(Board board, Integer sessionUserId) {
+            this(
+                    board.getId(),
+                    board.getTitle(),
+                    board.getContent(),
+                    board.getUser().getId(),
+                    board.getUser().getUsername(),
+                    sessionUserId.equals(board.getUser().getId()),
+                    board.getReplies().stream()
+                            .map(reply -> new ReplyDTO(reply, sessionUserId))
+                            .toList());
+        }
+
+        public record ReplyDTO(
+                Integer id,
+                String username,
+                String comment,
+                Boolean isOwner) {
+            public ReplyDTO(Reply reply, Integer sessionUserId) {
+                this(
+                        reply.getId(),
+                        reply.getUser().getUsername(),
+                        reply.getComment(),
+                        sessionUserId.equals(reply.getUser().getId()));
+            }
+        }
+    }
+}
