+++ src/main/java/com/metacoding/springv2/core/util/JWTUtil.java
@@ -1,11 +1,11 @@
 package com.metacoding.springv2.core.util;
 
+import java.util.Date;
+
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.metacoding.springv2.domain.user.User;
-
-import java.util.Date;
+import com.metacoding.springv2.user.User;
 
 public class JwtUtil {
     public static final String HEADER = "Authorization"; // HTTP 헤더 이름
