+++ src/main/java/com/metacoding/springv2/core/util/RespFilter.java
@@ -17,12 +17,16 @@ public class RespFilter {
         response.setContentType("application/json;charset=utf-8");
 
         Resp<?> resp = new Resp<>(status, msg, null);
+
         String responseBody = null;
         try {
             responseBody = om.writeValueAsString(resp);
         } catch (JsonProcessingException e) {
             log.error("JSON 변환 실패", e);
-            responseBody = "{\"status\":500, \"msg\":\"서버 내부 오류\"}";
+            responseBody = """
+                    {"status":500, "msg":"서버 내부 오류", body:null}
+                    """;
+            ;
         }
 
         PrintWriter out = response.getWriter();
