--- src/main/java/com/metacoding/springv2/domain/board/Board.java
@@ -1,47 +1,0 @@
-package com.metacoding.springv2.domain.board;
-
-import com.metacoding.springv2.domain.reply.Reply;
-import com.metacoding.springv2.domain.user.User;
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.sql.Timestamp;
-import java.util.*;
-
-@NoArgsConstructor
-@Getter
-@Entity
-@Table(name = "board_tb")
-public class Board {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    @Column(length = 30, nullable = false)
-    private String title;
-    @Column(length = 300, nullable = false)
-    private String content;
-
-    @CreationTimestamp
-    private Timestamp createdAt;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User user;
-
-    @OneToMany(mappedBy = "board", fetch = FetchType.LAZY)
-    private List<Reply> replies = new ArrayList<>();
-
-    public void update(String title, String content) {
-        this.title = title;
-        this.content = content;
-    }
-
-    @Builder
-    public Board(String title, String content, User user) {
-        this.title = title;
-        this.content = content;
-        this.user = user;
-    }
-}
