+++ src/main/java/com/metacoding/springv2/reply/Reply.java
@@ -1,0 +1,51 @@
+package com.metacoding.springv2.reply;
+
+import java.sql.Timestamp;
+
+import org.hibernate.annotations.CreationTimestamp;
+
+import com.metacoding.springv2.board.Board;
+import com.metacoding.springv2.user.User;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@Getter
+@Entity
+@Table(name = "reply_tb")
+public class Reply {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    @Column(length = 100, nullable = false)
+    private String comment;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Board board;
+
+    @CreationTimestamp
+    private Timestamp createdAt;
+
+    @Builder
+    public Reply(Integer id, String comment, User user, Board board, Timestamp createdAt) {
+        this.id = id;
+        this.comment = comment;
+        this.user = user;
+        this.board = board;
+        this.createdAt = createdAt;
+    }
+
+}
