+++ src/main/java/com/metacoding/springv2/reply/ReplyRepository.java
@@ -1,4 +1,4 @@
-package com.metacoding.springv2.domain.reply;
+package com.metacoding.springv2.reply;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
