+++ src/main/java/com/metacoding/springv2/reply/ReplyResponse.java
@@ -1,0 +1,24 @@
+package com.metacoding.springv2.reply;
+
+import com.metacoding.springv2.user.User;
+
+public class ReplyResponse {
+
+    public record DTO(
+            Integer id,
+            String comment,
+            Integer userId,
+            Integer boardId,
+            String username,
+            Boolean isOwner) {
+        public DTO(Reply reply, User sessionUser) {
+            this(
+                    reply.getId(),
+                    reply.getComment(),
+                    reply.getUser().getId(),
+                    reply.getBoard().getId(),
+                    sessionUser.getUsername(),
+                    reply.getUser().getId().equals(sessionUser.getId()));
+        }
+    }
+}
